[#2] 깃허브 소셜 로그인 API
- 패키지 구조 변경
