--- src/main/java/io/driver/codrive/domain/board/BoardRepository.java
@@ -1,8 +1,0 @@
-package io.driver.codrive.domain.board;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BoardRepository extends JpaRepository<Board, Long> {
-}
