--- src/main/java/io/driver/codrive/domain/follow/Follow.java
@@ -1,23 +1,0 @@
-package io.driver.codrive.domain.follow;
-
-import io.driver.codrive.domain.global.BaseEntity;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class Follow extends BaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long followId;
-
-	@Column(nullable = false)
-	private Boolean canceled;
-}
-
