--- src/main/java/io/driver/codrive/domain/follow/FollowRepository.java
@@ -1,8 +1,0 @@
-package io.driver.codrive.domain.follow;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface FollowRepository extends JpaRepository<Follow, Long> {
-}
