--- src/main/java/io/driver/codrive/domain/record/ProblemPlatform.java
@@ -1,5 +1,0 @@
-package io.driver.codrive.domain.record;
-
-public enum ProblemPlatform {
-	BAEKJOON, PROGRAMMERS, SWEA, LEETCODE, HACKERRANK, OTHER
-}
