--- src/main/java/io/driver/codrive/domain/record/RecordRepository.java
@@ -1,8 +1,0 @@
-package io.driver.codrive.domain.record;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RecordRepository extends JpaRepository<Record, Long> {
-}
