--- src/main/java/io/driver/codrive/domain/room/RoomRepository.java
@@ -1,8 +1,0 @@
-package io.driver.codrive.domain.room;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoomRepository extends JpaRepository<Room, Long> {
-}
