--- src/main/java/io/driver/codrive/domain/roomRequest/RoomRequestRepository.java
@@ -1,9 +1,0 @@
-package io.driver.codrive.domain.roomRequest;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-
-public interface RoomRequestRepository extends JpaRepository<RoomRequest, Long> {
-}
