--- src/main/java/io/driver/codrive/domain/tag/TagRepository.java
@@ -1,8 +1,0 @@
-package io.driver.codrive.domain.tag;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TagRepository extends JpaRepository<Tag, Long> {
-}
