--- src/main/java/io/driver/codrive/domain/tagRecordMap/TagRecordMap.java
@@ -1,19 +1,0 @@
-package io.driver.codrive.domain.tagRecordMap;
-
-import io.driver.codrive.domain.global.BaseEntity;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class TagRecordMap extends BaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long tagRecordMapId;
-}
