--- src/main/java/io/driver/codrive/domain/tagRecordMap/TagRecordMapRepository.java
@@ -1,8 +1,0 @@
-package io.driver.codrive.domain.tagRecordMap;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TagRecordMapRepository extends JpaRepository<TagRecordMap, Long> {
-}
