--- src/main/java/io/driver/codrive/domain/user/Language.java
@@ -1,5 +1,0 @@
-package io.driver.codrive.domain.user;
-
-public enum Language {
-	PYTHON, JAVA, C, CPP, CSHARP, JAVASCRIPT, GO, KOTLIN, RUBY, SWIFT, OTHER
-}
