--- src/main/java/io/driver/codrive/domain/user/Role.java
@@ -1,5 +1,0 @@
-package io.driver.codrive.domain.user;
-
-public enum Role {
-	OWNER, MEMBER
-}
