--- src/main/java/io/driver/codrive/domain/user/UserRepository.java
@@ -1,8 +1,0 @@
-package io.driver.codrive.domain.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-}
