--- src/main/java/io/driver/codrive/domain/userRoomMap/UserRoomMap.java
@@ -1,22 +1,0 @@
-package io.driver.codrive.domain.userRoomMap;
-
-import io.driver.codrive.domain.global.BaseEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserRoomMap extends BaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long userRoomMapId;
-}
