--- src/main/java/io/driver/codrive/domain/userRoomMap/UserRoomMapRepository.java
@@ -1,8 +1,0 @@
-package io.driver.codrive.domain.userRoomMap;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRoomMapRepository extends JpaRepository<UserRoomMap, Long> {
-}
