+++ src/main/java/io/driver/codrive/modules/board/domain/BoardRepository.java
@@ -1,0 +1,8 @@
+package io.driver.codrive.modules.board.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BoardRepository extends JpaRepository<Board, Long> {
+}
