+++ src/main/java/io/driver/codrive/modules/follow/domain/Follow.java
@@ -1,0 +1,23 @@
+package io.driver.codrive.modules.follow.domain;
+
+import io.driver.codrive.modules.global.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Follow extends BaseEntity {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long followId;
+
+	@Column(nullable = false)
+	private Boolean canceled;
+}
+
