+++ src/main/java/io/driver/codrive/modules/follow/domain/FollowRepository.java
@@ -1,0 +1,8 @@
+package io.driver.codrive.modules.follow.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface FollowRepository extends JpaRepository<Follow, Long> {
+}
