+++ src/main/java/io/driver/codrive/modules/record/domain/ProblemPlatform.java
@@ -1,0 +1,5 @@
+package io.driver.codrive.modules.record.domain;
+
+public enum ProblemPlatform {
+	BAEKJOON, PROGRAMMERS, SWEA, LEETCODE, HACKERRANK, OTHER
+}
