+++ src/main/java/io/driver/codrive/modules/record/domain/RecordRepository.java
@@ -1,0 +1,8 @@
+package io.driver.codrive.modules.record.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RecordRepository extends JpaRepository<Record, Long> {
+}
