+++ src/main/java/io/driver/codrive/modules/room/domain/RoomRepository.java
@@ -1,0 +1,8 @@
+package io.driver.codrive.modules.room.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoomRepository extends JpaRepository<Room, Long> {
+}
