+++ src/main/java/io/driver/codrive/modules/roomRequest/domain/RoomRequest.java
@@ -1,0 +1,23 @@
+package io.driver.codrive.modules.roomRequest.domain;
+
+import io.driver.codrive.modules.global.BaseEntity;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RoomRequest extends BaseEntity {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long roomRequestId;
+
+}
