+++ src/main/java/io/driver/codrive/modules/tag/domain/TagRepository.java
@@ -1,0 +1,8 @@
+package io.driver.codrive.modules.tag.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TagRepository extends JpaRepository<Tag, Long> {
+}
