+++ src/main/java/io/driver/codrive/modules/tagRecordMap/domain/TagRecordMap.java
@@ -1,0 +1,19 @@
+package io.driver.codrive.modules.tagRecordMap.domain;
+
+import io.driver.codrive.modules.global.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class TagRecordMap extends BaseEntity {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long tagRecordMapId;
+}
