+++ src/main/java/io/driver/codrive/modules/user/domain/Language.java
@@ -1,0 +1,5 @@
+package io.driver.codrive.modules.user.domain;
+
+public enum Language {
+	PYTHON, JAVA, C, CPP, CSHARP, JAVASCRIPT, GO, KOTLIN, RUBY, SWIFT, OTHER
+}
