+++ src/main/java/io/driver/codrive/modules/user/domain/Role.java
@@ -1,0 +1,5 @@
+package io.driver.codrive.modules.user.domain;
+
+public enum Role {
+	OWNER, MEMBER
+}
