+++ src/main/java/io/driver/codrive/modules/user/domain/User.java
@@ -1,0 +1,41 @@
+package io.driver.codrive.modules.user.domain;
+
+import io.driver.codrive.modules.global.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class User extends BaseEntity {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long userId;
+
+	@Column(nullable = false)
+	private String email;
+
+	@Column(nullable = false)
+	private String name;
+
+	@Column(nullable = false)
+	private String nickname;
+
+	private String comment;
+
+	@Column(nullable = false)
+	@Enumerated(EnumType.STRING)
+	private Language language;
+
+	@Column(nullable = false)
+	@Enumerated(EnumType.STRING)
+	private Role role;
+
+	@Column(nullable = false)
+	private Boolean withdraw;
+}
