+++ src/main/java/io/driver/codrive/modules/user/domain/UserRepository.java
@@ -1,0 +1,8 @@
+package io.driver.codrive.modules.user.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+}
