+++ src/main/java/io/driver/codrive/modules/userRoomMap/domain/UserRoomMapRepository.java
@@ -1,0 +1,8 @@
+package io.driver.codrive.modules.userRoomMap.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRoomMapRepository extends JpaRepository<UserRoomMap, Long> {
+}
